Fixed a problem where the attribute equation built-in COUNT() was misbehaving under certain circumstances.
git-svn-id: http://chianti.ucsd.edu/svn@19890 0ecc0d97-ab19-0410-9704-bfe1a75892f5

+++ cytoscape/branches/eqn_attribs/tests/cytoscape/data/eqn_attribs/interpreter/InterpreterTest.java
@@ -428,11 +428,11 @@ public class InterpreterTest extends TestCase {
 		numbers.add(5.0);
 		nameToDescriptorMap.put("list", new IdentDescriptor(numbers));
 		final Interpreter interpreter1 = new Interpreter(compiler.getEquation(), nameToDescriptorMap);
-		assertEquals(new Double(4.0), interpreter1.run());
+		assertEquals(new Long(4L), interpreter1.run());
 
 		assertTrue(compiler.compile("=COUNT(-2,-3,-4.35)", attribNameToTypeMap));
 		final Interpreter interpreter2 = new Interpreter(compiler.getEquation(), nameToDescriptorMap);
-		assertEquals(new Double(3.0), interpreter2.run());
+		assertEquals(new Long(3L), interpreter2.run());
 	}
 
 	public void testMEDIAN() throws Exception {
